Syncronize action state by getting json message from server
The Actions now get back a json response from the server that tells the
frontend the state it should be in after an action is called. Before,
when adding a todo the frontend would set the id. Now the backend
handles this and returns a json object to the frontend and the frontend
can add the fully qualified todo to the list.