Update testing for reverted AddTodo component
The original AddTodo component used a React `ref` property as well as an
HTML `<form/>` to set the state of the text to be passed to create a new
todo. This ened up being a better solution than implementing `onChange`
and `onSubmit` handles manually because the `<input/>` automatically
resets after the user clicks the add button.

Instead, the updated class component would keep the text around. We
could use `setState()`, but it's much easier to use `ref`.

The ReduxActions code was also updated to make it a bit more robust for
testing purposes. The test for each function was updated as well.