Update ReduxActions to give better error messages and change tests accordingly
Each Redux action used to give a generic eg: `Error: Bad response` but now they have
been updated to actually print out not only a better error message but the actual
status text from the server.

The tests for ReduxActions used to take an expected actions array and compare that
to the actual action that was received, but now it just checks if the action type was
a failure or whether it got the actuall success action back after the response.