Add loading animation to todo list when fetching from server
The ReduxApp component now shows a loading animation when trying to
fetch the initial state from the server and will display and error
message when the server cannot return state (such as the mongodb server
itself is down) or if the Flask server is unavailable.

Tests have been updated accordingly as well as the fact that the ReduxApp
now works a little differently. It now calls out an static async function
to help get state in `componentDidMount` that will set whether the app is
loading, errored, or successfull in getting the initial state.

Users can no longer try to input data into the todo system while the app
is loading or when the server is unavailable.